agents/walrus/requirements.txt
walrus-python>=0.1.0
requests>=2.0.0
python-dotenv>=1.0.0
uagents>=0.4.0
openai>=1.0.0 